// src/common/enums/errorType.enum.ts
export enum CustomErrorTypeEnum {
  DATA_NOT_FOUND = 1001,
  INCORRECT_CREDENTIALS = 1002,
  NO_ACCESS_OR_DATA_NOT_FOUND = 1003,
  ENTRY_ALREADY_EXISTS = 1004,
  ROW_POSITION_MORE_THAN_ALLOWED = 1005,
  SLIDER_NOT_ALLOWED_FOR_THIS_ROW = 1006, // row max col. should not be bigger than 1
  POSITION_ALREADY_EXISTS = 1007,
  LIMIT_FOR_ELEMENTS_IN_ROW = 1008,
  TYPE_OF_ELEMENT_NOT_ALLOWED = 1009,
  CAN_NOT_DELETE_LAST_SLIDE = 1010,
  INVALID_USER_CODE = 1011,
  USER_ALREADY_VERIFIED = 1012,
  TOO_MANY_REQUESTS = 1013,
  EXPERIMENT_ALREADY_HAS_THIS_STATUS = 1014,
  SLIDE_SHOULD_NOT_BE_PART_OF_CYCLE = 1015,
  ATTRIBUTES_LENGTH_CONFLICT = 1016, // attributes length and attributesName length did not match
  CAN_NOT_SET_TRAINING_FOR_CHILD_SLIDE = 1017,
  ARRAY_OF_ROW_DATA_NOT_UNIQUE = 1018,
  ATTRIBUTE_NOT_FOUND = 1019,
  VARIABLE_NOT_FOUND = 1020,
  ACCOUNT_NOT_VERIFIED = 1021,
  ID_NOT_FOUND = 1022,
  SLIDE_ID_NOT_FOUND = 1023,
  SLIDE_POSITION_NOT_FOUND = 1024,

  /* Experiment: start - GENERAL */
  EXPERIMENT_START_INCORRECT_USER_DATA = 1025,
  ACCESS_DENIED_FOR_THIS_PLATFORM = 1026,
  EXPERIMENT_START_CAN_NOT_START_UNPUBLISHED = 1027,

  /* Experiment: start - AGE */
  EXPERIMENT_START_AGE_EQUALS = 1028,
  EXPERIMENT_START_AGE_NOT_EQUALS = 1029,
  EXPERIMENT_START_AGE_MORE = 1030,
  EXPERIMENT_START_AGE_LESS = 1031,
  EXPERIMENT_START_AGE_MORE_OR_EQUALS = 1032,
  EXPERIMENT_START_AGE_LESS_OR_EQUALS = 1033,

  /* Experiment: start - SEX */
  EXPERIMENT_START_SEX_EQUALS = 1034,
  EXPERIMENT_START_SEX_NOT_EQUALS = 1035,

  /* Experiment: start - Native languages */
  EXPERIMENT_START_NATIVE_LANGUAGES_EQUALS = 1036,
  EXPERIMENT_START_NATIVE_LANGUAGES_NOT_EQUALS = 1037,

  /* Experiment: start - Learned languages */
  EXPERIMENT_START_LEARNED_LANGUAGES_EQUALS = 1038,
  EXPERIMENT_START_LEARNED_LANGUAGES_NOT_EQUALS = 1039,

  /* Variable */
  VARIABLE_CAN_NOT_SED_TO_SLIDE = 1040,
  VARIABLE_ARRAY_LENGTH_NOT_CORRECT = 1041,
  VARIABLE_NOT_ALL_COLUMNS_SET = 1042,
  VARIABLE_CAN_NOT_REMOVE_ALL_ATTRIBUTES = 1043,

  EXPERIMENT_ERROR_WHILE_CHECK_ACCESS = 1044,
  EXPERIMENT_CAN_NOT_UPDATE_PUBLISHED = 1045,

  BAD_REQUEST_IN_ENTITY = 1046,

  NO_DATA_FOR_CURRENT_SLIDE = 1047,
  NO_DATA_FOR_PREVIOUS_SLIDE = 1048,
  NO_DATA_FOR_START_TIME_PREVIOUS_SLIDE = 1049,
}